requirements.txt
streamlit
panda
numpy
plotly
altair
streamlit_pandas_profiling
